src/third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/v0/slotted-pseudo-element-in-v0v1-mixed-expected.txt

CONSOLE ERROR: line 10: Uncaught TypeError: v0host.createShadowRoot is not a function
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: v0host.createShadowRoot is not a function
Harness: the test ran to completion.
